--- src/services/reducers/userPassword/usePassword-interfaces.ts
@@ -1,9 +1,0 @@
-export const ADD_PASSWORD_REQUEST = '@user/addPasswordRequest';
-export const ADD_PASSWORD_SUCCESS = '@user/addPasswordSuccess';
-export const ADD_PASSWORD_ERROR = '@user/addPasswordError';
-
-export type FormValues = {
-  firstName: string;
-  lastName: string;
-  email: string;
-};
