+++ src/services/reducers/userPassword/usePassword-types.ts
@@ -1,0 +1,11 @@
+export type FormValues = {
+  firstName: string;
+  lastName: string;
+  email: string;
+};
+export type UserValues = {
+  firstName: string;
+  lastName: string;
+  email: string;
+  id: string;
+};
